@@ -19,7 +19,7 @@ class InternalsEventLoopManagerTests: XCTestCase {
         }
     }
 
-    private var manager: Internals.EventLoopGroupManager!
+    private var manager: Internals.EventLoopGroupManager?
 
     override func setUp() async throws {
         try await super.setUp()
@@ -36,7 +36,7 @@ class InternalsEventLoopManagerTests: XCTestCase {
         let provider = CustomProvider()
 
         // When
-        let sut1 = await manager.provider(provider)
+        let sut1 = try await XCTUnwrap(manager).provider(provider)
 
         // Then
         XCTAssertTrue(provider.group() === sut1)
@@ -47,8 +47,8 @@ class InternalsEventLoopManagerTests: XCTestCase {
         let provider = CustomProvider()
 
         // When
-        let sut1 = await manager.provider(provider)
-        let sut2 = await manager.provider(provider)
+        let sut1 = try await XCTUnwrap(manager).provider(provider)
+        let sut2 = try await XCTUnwrap(manager).provider(provider)
 
         // Then
         XCTAssertTrue(provider.group() === sut1)
@@ -60,8 +60,8 @@ class InternalsEventLoopManagerTests: XCTestCase {
         let provider = CustomProvider()
 
         // When
-        let sut = await _Concurrency.Task.detached(priority: .background) {
-            await self.manager.provider(provider)
+        let sut = try await _Concurrency.Task.detached(priority: .background) { [manager] in
+            try await XCTUnwrap(manager).provider(provider)
         }.value
 
         // Then
