@@ -7,22 +7,24 @@ import XCTest
 
 class ModifiersCollectBytesTests: XCTestCase {
 
-    var localServer: LocalServer!
+    var localServer: LocalServer?
 
     override func setUp() async throws {
         try await super.setUp()
         localServer = try await .init(.standard)
-        localServer.cleanup()
+        localServer?.cleanup()
     }
 
     override func tearDown() async throws {
         try await super.tearDown()
-        localServer.cleanup()
+        localServer?.cleanup()
         localServer = nil
     }
 
     func testCollect_whenUploadStep_shouldBeValid() async throws {
         // Given
+        let localServer = try XCTUnwrap(localServer)
+        
         let resource = Certificates().server()
         let message = "Hello World"
 
