@@ -7,22 +7,24 @@ import XCTest
 
 class ModifiersCollectDataTests: XCTestCase {
 
-    var localServer: LocalServer!
+    var localServer: LocalServer?
 
     override func setUp() async throws {
         try await super.setUp()
         localServer = try await .init(.standard)
-        localServer.cleanup()
+        localServer?.cleanup()
     }
 
     override func tearDown() async throws {
         try await super.tearDown()
-        localServer.cleanup()
+        localServer?.cleanup()
         localServer = nil
     }
 
     func testCollect_whenIsResultOfAsyncBytes() async throws {
         // Given
+        let localServer = try XCTUnwrap(localServer)
+
         let resource = Certificates().server()
         let message = "Hello World"
 
@@ -53,6 +55,8 @@ class ModifiersCollectDataTests: XCTestCase {
 
     func testCollect_whenIsAsyncBytes() async throws {
         // Given
+        let localServer = try XCTUnwrap(localServer)
+
         let resource = Certificates().server()
         let message = "Hello World"
 
@@ -83,6 +87,8 @@ class ModifiersCollectDataTests: XCTestCase {
 
     func testCollect_whenIsAsyncResponse() async throws {
         // Given
+        let localServer = try XCTUnwrap(localServer)
+
         let resource = Certificates().server()
         let message = "Hello World"
 
