@@ -7,22 +7,23 @@ import XCTest
 
 class UploadTaskTests: XCTestCase {
 
-    var localServer: LocalServer!
+    var localServer: LocalServer?
 
     override func setUp() async throws {
         try await super.setUp()
         localServer = try await .init(.standard)
-        localServer.cleanup()
+        localServer?.cleanup()
     }
 
     override func tearDown() async throws {
         try await super.tearDown()
-        localServer.cleanup()
+        localServer?.cleanup()
         localServer = nil
     }
 
     func testUploadTask() async throws {
         // Given
+        let localServer = try XCTUnwrap(localServer)
         let certificate = Certificates().server()
         let output = "Hello World"
         let upload = Data.randomData(length: 1_024)
